Move CallFrame to utils
Since this struct is used in multiple files, to avoid dependency cycles,
it is moved to utils.
